ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "onnx>=1.20.1",
    "onnxruntime>=1.24.4",
    "onnxscript>=0.6.2",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
